+++ packages/codegen/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @provablehq/veil-codegen
 
+## 0.8.0
+
+### Patch Changes
+
+- @provablehq/veil-core@0.8.0
+
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies [99defd6]
+  - @provablehq/veil-core@0.7.1
+
 ## 0.7.0
 
 ### Patch Changes