+++ packages/devnode/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @provablehq/veil-aleo-devnode
 
+## 0.8.0
+
+## 0.7.1
+
 ## 0.7.0
 
 ### Minor Changes