+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @provablehq/veil-aleo-react-hooks
 
+## 0.8.0
+
+### Patch Changes
+
+- @provablehq/veil-core@0.8.0
+- @provablehq/veil-aleo-wallet-adapter@0.8.0
+
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies [99defd6]
+  - @provablehq/veil-core@0.7.1
+  - @provablehq/veil-aleo-wallet-adapter@0.7.1
+
 ## 0.7.0
 
 ### Patch Changes