+++ packages/shield-swap-cli/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @provablehq/shield-swap-cli
 
+## 0.8.0
+
+### Patch Changes
+
+- Updated dependencies [cbb80ef]
+  - @provablehq/shield-swap-sdk@0.8.0
+  - @provablehq/veil-aleo-sdk@0.8.0
+
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [99defd6]
+  - @provablehq/veil-aleo-sdk@0.7.1
+  - @provablehq/shield-swap-sdk@0.7.1
+
 ## 0.7.0
 
 ### Minor Changes