+++ packages/wallet-adapter/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @provablehq/veil-aleo-wallet-adapter
 
+## 0.8.0
+
+## 0.7.1
+
 ## 0.7.0
 
 ## 0.6.0